+++ pages/form-builder/responses/[[...params]].tsx
@@ -7,7 +7,7 @@ import { authOptions } from "@pages/api/auth/[...nextauth]";
 import { AccessControlError, createAbility } from "@lib/privileges";
 import { NextPageWithLayout } from "@pages/_app";
 import { PageTemplate, Template } from "@components/form-builder/app";
-import { Button, useDialogRef, Dialog, LineItemEntries } from "@components/form-builder/app/shared";
+import { Button } from "@components/form-builder/app/shared";
 import { StyledLink } from "@components/globals/StyledLink/StyledLink";
 import { GetServerSideProps } from "next";
 import { FormRecord, VaultSubmissionList } from "@lib/types";
@@ -17,7 +17,8 @@ import { useSession } from "next-auth/react";
 import Head from "next/head";
 import { checkOne } from "@lib/cache/flags";
 import Link from "next/link";
-import { isUUID, isFormId } from "@lib/validation";
+import { DialogConfirmReceipt } from "./DialogConfirmReceipt";
+import { DialogReportProblems } from "./DialogReportProblems";
 
 interface ResponsesProps {
   vaultSubmissions: VaultSubmissionList[];
@@ -27,43 +28,11 @@ const Responses: NextPageWithLayout<ResponsesProps> = ({ vaultSubmissions }: Res
   const { t } = useTranslation("form-builder");
   const { status } = useSession();
   const isAuthenticated = status === "authenticated";
+  const [isShowConfirmReceiptDialog, setIsShowConfirmReceiptDialog] = useState(false);
+  const [isShowReportProblemsDialog, setIsShowReportProblemsDialog] = useState(false);
 
   const secondaryButtonClass =
     "whitespace-nowrap text-sm rounded-full bg-white-default text-black-default border-black-default hover:text-white-default hover:bg-gray-600 active:text-white-default active:bg-gray-500 py-2 px-5 rounded-lg border-2 border-solid inline-flex items-center active:top-0.5 focus:outline-[3px] focus:outline-blue-focus focus:outline focus:outline-offset-2 focus:bg-blue-focus focus:text-white-default disabled:cursor-not-allowed disabled:text-gray-500";
-
-  // Dialog: Confirm Code vars
-  const [codes, setCodes] = useState<string[]>([]);
-  const validateCodes = (code: string) => {
-    return isUUID(code);
-  };
-  const dialogConfirmReceipt = useDialogRef();
-  const dialogConfirmReceiptInstructionId =
-    "dialog-confirm-receipt-instruction-" + Math.random().toString(36).substr(2, 9);
-  const [isShowConfirmReceiptDialog, setIsShowConfirmReceiptDialog] = useState(false);
-  const dialogConfirmReceiptHandleClose = () => {
-    setIsShowConfirmReceiptDialog(false);
-    dialogConfirmReceipt.current?.close();
-  };
-  const handleConfirmReceiptSubmit = () => {
-    //TODO
-  };
-
-  // Dialog: Report Problems vars
-  const [formNumbers, setFormNumbers] = useState<string[]>([]);
-  const validateFormNumber = (formId: string) => {
-    return isFormId(formId);
-  };
-  const dialogReportProblems = useDialogRef();
-  const dialogReportProblemsInstructionId =
-    "dialog-report-problems-instruction-" + Math.random().toString(36).substr(2, 9);
-  const [isShowReportProblemsDialog, setIsShowReportProblemsDialog] = useState(false);
-  const dialogReportProblemsHandleClose = () => {
-    setIsShowReportProblemsDialog(false);
-    dialogReportProblems.current?.close();
-  };
-  const handleReportProblemSubmit = () => {
-    //TODO
-  };
 
   return (
     <>
@@ -177,69 +146,15 @@ const Responses: NextPageWithLayout<ResponsesProps> = ({ vaultSubmissions }: Res
         )}
       </PageTemplate>
 
-      {isShowConfirmReceiptDialog && (
-        <Dialog
-          title={t("responses.confirmReceiptDialog.title")}
-          dialogRef={dialogConfirmReceipt}
-          handleClose={dialogConfirmReceiptHandleClose}
-          headerStyle="inline-block ml-12 mt-12"
-        >
-          <div className="px-10 py-4">
-            <p className="mb-8">{t("responses.confirmReceiptDialog.findCode")}</p>
-            <p className="mt-20 mb-2 font-bold" id={dialogConfirmReceiptInstructionId}>
-              {t("responses.confirmReceiptDialog.copyCode")}
-            </p>
-            <LineItemEntries
-              inputs={codes}
-              setInputs={setCodes}
-              validateInput={validateCodes}
-              spellCheck={false}
-              inputLabelId={dialogConfirmReceiptInstructionId}
-            ></LineItemEntries>
-            <p className="mt-8">{t("responses.confirmReceiptDialog.responsesAvailableFor")}</p>
-            <div className="flex mt-8 mb-8">
-              <Button className="mr-4" onClick={handleConfirmReceiptSubmit}>
-                {t("responses.confirmReceipt")}
-              </Button>
-              <Button theme="secondary" onClick={dialogConfirmReceiptHandleClose}>
-                {t("cancel")}
-              </Button>
-            </div>
-          </div>
-        </Dialog>
-      )}
+      <DialogConfirmReceipt
+        isShowDialog={isShowConfirmReceiptDialog}
+        setIsShowDialog={setIsShowConfirmReceiptDialog}
+      />
 
-      {isShowReportProblemsDialog && (
-        <Dialog
-          title={t("responses.reportProblemsDialog.title")}
-          dialogRef={dialogReportProblems}
-          handleClose={dialogReportProblemsHandleClose}
-          headerStyle="inline-block ml-12 mt-12"
-        >
-          <div className="px-10 py-4">
-            <p className="mb-8">{t("responses.reportProblemsDialog.findForm")}</p>
-            <p id={dialogReportProblemsInstructionId} className="mt-20 mb-2 font-bold">
-              {t("responses.reportProblemsDialog.enterFormNumbers")}
-            </p>
-            <LineItemEntries
-              inputs={formNumbers}
-              setInputs={setFormNumbers}
-              validateInput={validateFormNumber}
-              spellCheck={false}
-              inputLabelId={dialogReportProblemsInstructionId}
-            ></LineItemEntries>
-            <p className="mt-8">{t("responses.reportProblemsDialog.problemReported")}</p>
-            <div className="flex mt-8 mb-8">
-              <Button className="mr-4" onClick={handleReportProblemSubmit}>
-                {t("responses.reportProblemsDialog.reportProblems")}
-              </Button>
-              <Button theme="secondary" onClick={dialogReportProblemsHandleClose}>
-                {t("cancel")}
-              </Button>
-            </div>
-          </div>
-        </Dialog>
-      )}
+      <DialogReportProblems
+        isShowDialog={isShowReportProblemsDialog}
+        setIsShowDialog={setIsShowReportProblemsDialog}
+      />
     </>
   );
 };
